Fix test

+++ test/test_utils.py
@@ -318,8 +318,8 @@ def test_draw_keypoints_errors():
 
 
 def test_flow_to_image_errors():
-    wrong_flow1 = torch.full((3, 10, 10), 0)
-    wrong_flow2 = torch.full((2, 10), 0)
+    wrong_flow1 = torch.full((3, 10, 10), 0, dtype=torch.float)
+    wrong_flow2 = torch.full((2, 10), 0, dtype=torch.float)
     wrong_flow3 = torch.full((2, 10, 30), 0, dtype=torch.int)
 
     with pytest.raises(ValueError, match="Input flow should have shape"):
